Stage to Prod Sync

+++ docroot/profiles/unicef/modules/custom/vss_custom/src/Controller/ExitController.php
@@ -2,14 +2,14 @@
 
 namespace Drupal\vss_custom\Controller;
 
+use Drupal\Core\State\State;
+use Drupal\Core\Session\SessionManager;
 use Drupal\Core\Controller\ControllerBase;
 use Drupal\Core\Routing\TrustedRedirectResponse;
-use Drupal\Core\DependencyInjection\ContainerInjectionInterface;
-use Symfony\Component\DependencyInjection\ContainerInterface;
-use Drupal\Core\Session\SessionManager;
-use Drupal\Core\State\State;
-use Drupal\Core\PageCache\ResponsePolicy\KillSwitch;
 use Symfony\Component\HttpFoundation\RequestStack;
+use Drupal\Core\PageCache\ResponsePolicy\KillSwitch;
+use Symfony\Component\DependencyInjection\ContainerInterface;
+use Drupal\Core\DependencyInjection\ContainerInjectionInterface;
 
 /**
  * Class ExitController.
@@ -43,7 +43,12 @@ class ExitController extends ControllerBase implements ContainerInjectionInterfa
   /**
    * ExitController constructor.
    */
-  public function __construct(SessionManager $session_manager, State $state, KillSwitch $page_cache_kill_switch, RequestStack $request_stack) {
+  public function __construct(
+    SessionManager $session_manager,
+    State $state,
+    KillSwitch $page_cache_kill_switch,
+    RequestStack $request_stack) {
+
     $this->sessionManager = $session_manager;
     $this->state = $state;
     $this->pageCacheKillSwitch = $page_cache_kill_switch;
@@ -67,14 +72,19 @@ class ExitController extends ControllerBase implements ContainerInjectionInterfa
    */
   public function clearAllData() {
     $exit_url = $this->state->get('exit_url');
-    $this->sessionManager->destroy();
-    if (isset($_SERVER['HTTP_COOKIE'])) {
-      $cookies = explode(';', $_SERVER['HTTP_COOKIE']);
-      foreach ($cookies as $cookie) {
-        $parts = explode('=', $cookie);
-        $name = trim($parts[0]);
-        setcookie($name, '', time() - 1000);
-        setcookie($name, '', time() - 1000, '/');
+    if ($this->sessionManager->isStarted()) {
+      $this->sessionManager->destroy();
+    }
+
+    $cookies = $this->requestStack->cookies->all();
+    foreach ($cookies as $name => $cookie) {
+      setcookie($name, '', time() - 1000);
+      setcookie($name, '', time() - 1000, '/');
+      if ($name == 'country-location-selector') {
+        $domain_current_url = explode(".", $this->requestStack->server->get('SERVER_NAME'));
+        $domain_slice = array_slice($domain_current_url, -2);
+        $domain_site = '.' . $domain_slice[0] . '.' . $domain_slice[1];
+        setcookie($name, '', time() - 100, "/", $domain_site);
       }
     }
     $this->pageCacheKillSwitch->trigger();
